test(unity-project): split resolution tests by class responsibility
Why:
- DefaultUnityEditorSearchRootProviderTests に provider 以外の責務テストが混在し、保守単位が粗くなっていたため。

What Changed:
- provider 専用ケースのみを DefaultUnityEditorSearchRootProviderTests に残した。
- UnityEditorSearchRootBuilderTests を新設した。
- UnityPathComparerProviderTests を新設した。
- Linux/Mac source の検証をそれぞれ専用テストファイルへ分離した。

Impact:
- クラス責務とテスト責務が一致し、変更時の差分ノイズと原因切り分けコストを削減する。
