+++ tests/Ucli.Tests/Unity/Resolution/UnityEditorSearchRootBuilderTests.cs
@@ -1,0 +1,17 @@
+namespace MackySoft.Ucli.Tests;
+
+using MackySoft.Ucli.UnityProject.Resolution;
+
+public sealed class UnityEditorSearchRootBuilderTests
+{
+    [Fact]
+    [Trait("Size", "Small")]
+    public void Add_WhenRootPathIsNullOrWhitespace_ThrowsArgumentException ()
+    {
+        var builder = new UnityEditorSearchRootBuilder(StringComparer.Ordinal);
+
+        Assert.ThrowsAny<ArgumentException>(() => builder.Add(null));
+        Assert.ThrowsAny<ArgumentException>(() => builder.Add(string.Empty));
+        Assert.ThrowsAny<ArgumentException>(() => builder.Add(" "));
+    }
+}
