+++ tests/Ucli.Tests/Unity/Resolution/UnityPathComparerProviderTests.cs
@@ -1,0 +1,17 @@
+namespace MackySoft.Ucli.Tests;
+
+using MackySoft.Ucli.UnityProject.Resolution;
+
+public sealed class UnityPathComparerProviderTests
+{
+    [Fact]
+    [Trait("Size", "Small")]
+    public void GetComparer_ReturnsPlatformAwareComparer ()
+    {
+        var provider = new UnityPathComparerProvider();
+
+        var comparer = provider.GetComparer();
+
+        Assert.Equal(OperatingSystem.IsWindows(), comparer.Equals("A", "a"));
+    }
+}
